Fixed: Restructured the customer registration API from POST /auth/register to POST /customers, making it RESTful
