+++ packages/next/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@spree/next",
-  "version": "0.7.2",
+  "version": "0.8.0",
   "description": "Next.js integration for Spree Commerce — server actions, caching, and cookie-based auth",
   "type": "module",
   "main": "./dist/index.js",
@@ -62,7 +62,7 @@
     "url": "https://github.com/spree/spree/issues"
   },
   "peerDependencies": {
-    "@spree/sdk": ">=0.7.2",
+    "@spree/sdk": ">=0.8.0",
     "next": ">=15.0.0",
     "react": ">=19.0.0"
   },
