+++ spree/api/spec/integration/spree/api/v3/store/auth_spec.rb
@@ -76,93 +76,6 @@ RSpec.describe 'Authentication API', type: :request, swagger_doc: 'api-reference
     end
   end
 
-  path '/api/v3/store/auth/register' do
-    post 'Register' do
-      tags 'Authentication'
-      consumes 'application/json'
-      produces 'application/json'
-      security [api_key: []]
-      description 'Creates a new customer account and returns a JWT token'
-
-      sdk_example <<~JS
-        const auth = await client.auth.register({
-          email: 'newuser@example.com',
-          password: 'password123',
-          password_confirmation: 'password123',
-          first_name: 'John',
-          last_name: 'Doe',
-          phone: '+1234567890',
-          accepts_email_marketing: true,
-          metadata: { source: 'storefront' },
-        })
-      JS
-
-      parameter name: 'x-spree-api-key', in: :header, type: :string, required: true
-      parameter name: :body, in: :body, schema: {
-        type: :object,
-        properties: {
-          email: { type: :string, format: 'email', example: 'newuser@example.com' },
-          password: { type: :string, minLength: 6, example: 'password123' },
-          password_confirmation: { type: :string, example: 'password123' },
-          first_name: { type: :string, example: 'John' },
-          last_name: { type: :string, example: 'Doe' },
-          phone: { type: :string, example: '+1234567890' },
-          accepts_email_marketing: { type: :boolean, example: true },
-          metadata: { type: :object, example: { source: 'storefront' } }
-        },
-        required: %w[email password]
-      }
-
-      response '201', 'registration successful' do
-        let(:'x-spree-api-key') { api_key.token }
-        let(:body) do
-          {
-            email: 'newuser@example.com',
-            password: 'password123',
-            first_name: 'John',
-            last_name: 'Doe',
-            phone: '+1234567890',
-            accepts_email_marketing: true,
-            metadata: { source: 'storefront' }
-          }
-        end
-
-        schema '$ref' => '#/components/schemas/AuthResponse'
-
-        run_test! do |response|
-          data = JSON.parse(response.body)
-          expect(data['token']).to be_present
-          expect(data['user']).to be_present
-
-          # Verify user was created with all fields
-          new_user = Spree.user_class.find_by(email: 'newuser@example.com')
-          expect(new_user).to be_present
-          expect(new_user.first_name).to eq('John')
-          expect(new_user.phone).to eq('+1234567890')
-          expect(new_user.accepts_email_marketing).to eq(true)
-          expect(new_user.metadata).to eq({ 'source' => 'storefront' })
-        end
-      end
-
-      response '422', 'email already taken' do
-        let(:'x-spree-api-key') { api_key.token }
-        let(:body) do
-          {
-            email: existing_user.email,
-            password: 'password123'
-          }
-        end
-
-        schema '$ref' => '#/components/schemas/ErrorResponse'
-
-        run_test! do |response|
-          data = JSON.parse(response.body)
-          expect(data['error']['code']).to eq('validation_error')
-        end
-      end
-    end
-  end
-
   path '/api/v3/store/auth/refresh' do
     post 'Refresh token' do
       tags 'Authentication'
